Move Prettifier into its own module.
to solve a cyclic dependency.
