+++ cypher/cypher-prettifier/src/test/scala/org/neo4j/cypher/docgen/tooling/ParserTestBase.scala
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.docgen.tooling.tests
+package org.neo4j.cypher.docgen.tooling
 
 import org.neo4j.cypher.GraphIcing
 import org.opencypher.v9_0.parser.BufferPosition
