+++ cypher/graphgist/pom.xml
@@ -29,12 +29,12 @@
         <!-- neo4j-documentation -->
         <dependency>
             <groupId>org.neo4j.doc</groupId>
-            <artifactId>neo4j-cypher-docs</artifactId>
+            <artifactId>test-framework-enterprise</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.neo4j.doc</groupId>
-            <artifactId>test-framework-enterprise</artifactId>
+            <artifactId>neo4j-cypher-prettifier</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
