+++ cypher/pom.xml
@@ -24,6 +24,7 @@
 
     <modules>
         <module>cypher-docs</module>
+        <module>cypher-prettifier</module>
         <module>graphgist</module>
         <module>refcard-tests</module>
     </modules>
