+++ cypher/refcard-tests/pom.xml
@@ -89,6 +89,11 @@
         </dependency>
         <dependency>
             <groupId>org.neo4j.doc</groupId>
+            <artifactId>neo4j-cypher-prettifier</artifactId>
+            <version>${neo4j.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.neo4j.doc</groupId>
             <artifactId>neo4j-graphviz</artifactId>
             <version>${neo4j.version}</version>
             <scope>test</scope>
