+++ embedded-examples/pom.xml
@@ -73,7 +73,7 @@
         </dependency>
         <dependency>
             <groupId>org.neo4j.doc</groupId>
-            <artifactId>neo4j-cypher-docs</artifactId>
+            <artifactId>neo4j-cypher-prettifier</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
