+++ server-test-utils/pom.xml
@@ -25,7 +25,7 @@
         </dependency>
         <dependency>
             <groupId>org.neo4j.doc</groupId>
-            <artifactId>neo4j-cypher-docs</artifactId>
+            <artifactId>neo4j-cypher-prettifier</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
